pipeline/requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2
shapely>=2.0.0
pandas>=2.0.0
snowflake-connector-python[pandas]>=3.5.0
